renames stage_time_scheme to NSTAGE_TIME_SCHEME (as in 3D codes); adds local parameter to output backward wavefield traces (for debugging)

+++ EXAMPLES/poroelastic_semi_infinite_homogeneous/run_this_example.sh
@@ -15,7 +15,11 @@ echo
 mkdir -p OUTPUT_FILES
 
 # cleans output files
-rm -rf OUTPUT_FILES/*
+if [ "$1" != "noclean" ]; then
+  echo "cleaning OUTPUT_FILES/"
+  # cleans output files
+  rm -rf OUTPUT_FILES/*
+fi
 
 cd $currentdir
